fix(nika-verb-infer): a paid blank answer fails typed (NIKA-INFER-004)
A thinking model under a tight `max_tokens` can spend the whole budget
on its reasoning trace and conclude with a BLANK visible answer — the
task settled green over `""` (the OBS-E warn) while every downstream
`${{ tasks.X.output }}` silently resolved to nothing (#651 · the paid
repro: 7/7 done · exit 0 · an empty brief.md).

- the warn is promoted to the typed `NIKA-INFER-004` failure (engine
  NIKA-435), fail-closed, the spend attached as ledger evidence
  (`SpendOnFailure`) — the registry, the wire mapping, the vendored
  canon row and the spec 05 table all name it
- the signal is narrow on purpose: blank visible answer PAIRED with
  token spend (a reported reasoning split, or an undifferentiated
  `output_tokens > 0` — ollama strips the think block upstream). Zero
  spend stays green; the `schema:` lane is untouched (an empty reply
  already dies NIKA-INFER-002 at extraction); `on_error:` stays the
  author's named opt-out
- the « N/N done » summary can no longer count a paid empty answer as
  done — the failure default makes the honesty leg moot
- nika-verb-infer's lib tests split to `tests.rs`/`proptests.rs`
  (the 15k wall), no behavior change there

Closes #651
