+++ estate.yaml
@@ -18,11 +18,11 @@ classes:
   testimonial: "captured evidence — fixtures, corpora, traces, receipts: proves behavior; neither shipped source nor regenerable output"
   foreign: "pointer to a third-party sovereign source"
 summary:
-  classified_files: 4370
+  classified_files: 4372
   file_rows: 11
   pattern_rows: 22
   by_class:
-    authored: 1439
+    authored: 1441
     authored-pin: 2
     generated: 118
     pinned-copy: 115
@@ -37,7 +37,7 @@ files:
   note: "carries the '<!-- AUTO-GENERATED by scripts/refresh-status.sh — do not edit by hand -->' status block (vector 23 parity-enforced) — a tool-maintained block inside an authored file"
 - path: "CHANGELOG.md"
   class: generated
-  sha256: 9c5a3919d99abb61b19c4c82eee85f7101630f4ade9920b498872026e17af6d8
+  sha256: 41c1d97c1293506a42d9feb7c34a68463764f1e2bfbdf2714c51876531331b5c
   evidence: "changelog-cliff.yml commits 'docs(changelog): append <tag> — auto-generated' on tag push (git-cliff), branch chore/changelog-<tag>, PR body 'Review, edit if needed, merge'"
   derivation:
     tool: "git-cliff (cliff.toml) via .github/workflows/changelog-cliff.yml on tag push"
@@ -114,7 +114,7 @@ patterns:
 - glob: "crates/nika-pack/pack/**"
   class: pinned-copy
   match_count: 114
-  aggregate_sha256: aed224dfaf83990a8648c3a984829c1cfd0fcba50a2b011255e81fed35a24533
+  aggregate_sha256: d6862483be6551307976ca1686a32fec550a0a7a444c1a3e785e7ebaddcac4d1
   evidence: "crates/nika-pack/src/lib.rs:16 'The snapshot under pack/ is vendored by scripts/sync-pack.sh' — a byte-copy of the spec surface, never edited here"
   derivation:
     tool: "bash scripts/sync-pack.sh <nika-spec checkout> — healed daily by .github/workflows/pack-resync.yml (branch bot/pack-resync · PR-only, the 12-gate CI judges)"
@@ -151,8 +151,8 @@ patterns:
     - "cargo-public-api 0.51.0"
 - glob: "crates/**/src/**"
   class: authored
-  match_count: 724
-  aggregate_sha256: 59291f433f4cc930a73a57f6b26a903effe96f0618b16cd470506911af95c94b
+  match_count: 726
+  aggregate_sha256: cf13fb403a980ae515af57f18a78618c8066e99959ad92f91c18b5444ab6cfea
   evidence: "the code — SPDX AGPL-3.0-or-later headers · 12-gate admission (ADR-003) · zero-unwrap discipline; spot-checked file heads carry no generation marker (grep hits for 'generated' are prose mentions, not markers)"
 - glob: "crates/**/tests/**"
   class: authored
@@ -178,7 +178,7 @@ patterns:
 - glob: "docs/**"
   class: authored
   match_count: 101
-  aggregate_sha256: ec474805ccf06bd054086f1a97c0412ba1c6232a91ed3ba5fad12cd49a4d500d
+  aggregate_sha256: 5d73a1f1a3f378efa6ab78bed90d2ea8585cff785ac9c0d8dc7fd00a0af380be
   evidence: "crate-specs · architecture prose · perf notes — no generation markers at file heads"
 - glob: "scripts/**"
   class: authored
